+++ android/koin-android/src/main/java/org/koin/android/ext/android/AndroidKoinScopeExt.kt
@@ -1,5 +1,6 @@
-package org.koin.androidx.viewmodel.ext.android
+package org.koin.android.ext.android
 
+import android.content.ComponentCallbacks
 import org.koin.android.scope.AndroidScopeComponent
 import org.koin.core.annotation.KoinInternalApi
 import org.koin.core.component.KoinComponent
@@ -10,11 +11,11 @@ import org.koin.core.scope.Scope
 
 @KoinInternalApi
 @PublishedApi
-internal fun getKoinScope(any : Any): Scope {
-    return when (any) {
-        is AndroidScopeComponent -> any.scope
-        is KoinScopeComponent -> any.scope
-        is KoinComponent -> any.getKoin().scopeRegistry.rootScope
+internal fun ComponentCallbacks.getKoinScope(): Scope {
+    return when (this) {
+        is AndroidScopeComponent -> scope
+        is KoinScopeComponent -> scope
+        is KoinComponent -> getKoin().scopeRegistry.rootScope
         else -> GlobalContext.get().scopeRegistry.rootScope
     }
 }