+++ android/koin-android/src/main/java/org/koin/androidx/viewmodel/ext/android/SharedSavedStateRegistryOwnerExt.kt
@@ -18,6 +18,7 @@ package org.koin.androidx.viewmodel.ext.android
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelStoreOwner
+import org.koin.android.ext.android.getKoinScope
 import org.koin.androidx.viewmodel.ViewModelOwner.Companion.from
 import org.koin.androidx.viewmodel.scope.BundleDefinition
 import org.koin.androidx.viewmodel.scope.emptyState
@@ -32,8 +33,7 @@ import kotlin.reflect.KClass
  *
  * @author Arnaud Giuliani
  */
-@Deprecated("getStateViewModel will be merged to sharedViewModel",
-    ReplaceWith("sharedViewModel(qualifier,parameters = parameters)"))
+@Deprecated("getStateViewModel will be merged to sharedViewModel - no need anymore of state parameter")
 inline fun <reified T : ViewModel> Fragment.sharedStateViewModel(
     qualifier: Qualifier? = null,
     noinline state: BundleDefinition = emptyState(),
@@ -44,8 +44,7 @@ inline fun <reified T : ViewModel> Fragment.sharedStateViewModel(
     }
 }
 
-@Deprecated("getStateViewModel will be merged to sharedViewModel",
-    ReplaceWith("sharedViewModel(qualifier,clazz,parameters = parameters)"))
+@Deprecated("getStateViewModel will be merged to sharedViewModel - no need anymore of state parameter")
 fun <T : ViewModel> Fragment.sharedStateViewModel(
     qualifier: Qualifier? = null,
     state: BundleDefinition = emptyState(),
@@ -55,8 +54,7 @@ fun <T : ViewModel> Fragment.sharedStateViewModel(
     return lazy(LazyThreadSafetyMode.NONE) { getStateViewModel(qualifier, state, clazz, parameters) }
 }
 
-@Deprecated("getStateViewModel will be merged to getViewModel",
-    ReplaceWith("getViewModel(qualifier,parameters = parameters)"))
+@Deprecated("getStateViewModel will be merged to getViewModel - no need anymore of state parameter")
 inline fun <reified T : ViewModel> Fragment.getStateViewModel(
     qualifier: Qualifier? = null,
     noinline state: BundleDefinition = emptyState(),
@@ -65,8 +63,7 @@ inline fun <reified T : ViewModel> Fragment.getStateViewModel(
     return getStateViewModel(qualifier, state, T::class, parameters)
 }
 
-@Deprecated("getStateViewModel will be merged to getViewModel",
-    ReplaceWith("getViewModel(qualifier,clazz,parameters = parameters)"))
+@Deprecated("getStateViewModel will be merged to getViewModel - no need anymore of state parameter")
 @OptIn(KoinInternalApi::class)
 fun <T : ViewModel> Fragment.getStateViewModel(
     qualifier: Qualifier? = null,
@@ -74,5 +71,5 @@ fun <T : ViewModel> Fragment.getStateViewModel(
     clazz: KClass<T>,
     parameters: ParametersDefinition? = null,
 ): T {
-    return getKoinScope(this).getViewModel(qualifier, state, { from(activity as ViewModelStoreOwner, this) }, clazz, parameters)
+    return getKoinScope().getViewModel(qualifier, state, { from(activity as ViewModelStoreOwner, this) }, clazz, parameters)
 }